Algorithm 1 file to server-side
